removed co.join() from all examples
